# Swagger-export/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "MDK Mashup API",
        "description": "Hämta trailer och information om alla världens filmer.",
        "version": "1.0.0"
    },
    "host": "maxtax.pythonanywhere.com",
    "schemes": [
        "https"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/movie/{imdbID}": {
            "get": {
                "summary": "Filmdata",
                "description": "Hämtar information om given film baserat på deras IMDB-id. Bland annat hämtas information om skådespelare, författare och trailer.  \n",
                "parameters": [
                    {
                        "name": "imdbID",
                        "in": "path",
                        "description": "En films unika ID. Exempel: tt2310332, tt0076759",
                        "required": true,
                        "type": "string",
                        "format": "single"
                    }
                ],
                "tags": [
                    "Movie"
                ],
                "responses": {
                    "200": {
                        "description": "En lista med filmdata.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Movie"
                            }
                        }
                    },
                    "default": {
                        "description": "Incorrect IMDb ID",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Error"
                            }
                        }
                    }
                }
            }
        },
        "/search?q={search}": {
            "get": {
                "summary": "Filmdata",
                "description": "Hämtar information om filmer baserat på söksträng.",
                "parameters": [
                    {
                        "name": "search",
                        "in": "path",
                        "description": "En söksträng",
                        "required": true,
                        "type": "string",
                        "format": "single"
                    }
                ],
                "tags": [
                    "Search"
                ],
                "responses": {
                    "200": {
                        "description": "En lista med sökdata.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Search"
                            }
                        }
                    },
                    "default": {
                        "description": "Movie not found!",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Error"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Movie": {
            "type": "object",
            "properties": {
                "Actors": {
                    "type": "string",
                    "description": "En sträng med samtliga skådespelare avgränsat med ett komma."
                },
                "Plot": {
                    "type": "string",
                    "description": "Filmens handling"
                },
                "Title": {
                    "type": "string",
                    "description": "Filmens titel."
                },
                "Writer": {
                    "type": "string",
                    "description": "Filmens författare avgränsat med komma"
                },
                "imdbID": {
                    "type": "string",
                    "description": "En films unika ID."
                },
                "trailer_embed": {
                    "type": "string",
                    "description": "Iframe med trailer."
                },
                "trailer_link": {
                    "type": "string",
                    "description": "Länk till trailer på traileraddict."
                }
            }
        },
        "Search": {
            "type": "object",
            "properties": {
                "Title": {
                    "type": "string",
                    "description": "Filmens titel."
                },
                "Year": {
                    "type": "string",
                    "description": "Året då filmen släpptes."
                },
                "imdbID": {
                    "type": "string",
                    "description": "En films unika ID."
                }
            }
        },
        "Error": {
            "type": "object",
            "properties": {
                "Error": {
                    "type": "string"
                },
                "Response": {
                    "type": "string"
                }
            }
        }
    }
}